backend/requirements.txt
fastapi[standard]
sqlalchemy
psycopg[async]
alembic
email-validator
pyjwt
passlib[bcrypt]